drop the useless cast on OTUs.size() (bugprone-useless-cast)
OTU_map::size_type is std::size_t, and OTU::input_order is already
unsigned long int, so on the platforms mumu targets the cast converts a
type to itself. It broke `make debug`, which compiles with -Werror and
-Wuseless-cast, and had done so since 79df690.

Removing the cast rather than silencing the warning is the right fix
here: -Wconversion and -Wsign-conversion are also enabled in that build
and stay quiet, which is the compiler confirming the two types match
exactly. Should that ever stop being true, the assignment becomes a
narrowing conversion and the debug build will say so, which is more
useful than a cast that hides it.

`make debug` (-O0 -D_GLIBCXX_DEBUG -fsanitize=undefined,address -Werror)
compiles clean again and passes 209/209 tests.

+++ src/load_OTUs.cpp
@@ -119,7 +119,7 @@ namespace {
     // get abundance values (rest of the line, we know there are n samples)
     // the map scrambles order; each line inserts exactly one OTU (duplicates
     // are fatal above), so the running entry count is this OTU's input order
-    otu.input_order = static_cast<unsigned long int>(OTUs.size());
+    otu.input_order = OTUs.size();
     otu.samples.reserve(n_samples);
     auto remaining = std::string_view{line}.substr(first_sep + 1);
     while (not remaining.empty()) {
